Module Make.Contract_repr

type t = private

| Implicit of Tezos_protocol_environment.Signature.Public_key_hash.t

| Originated of Contract_hash.t

type contract = t

include Tezos_protocol_environment.Compare.S with type t := contract

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

Implicit contracts

val implicit_contract : Tezos_protocol_environment.Signature.Public_key_hash.t -> contract

val is_implicit : contract -> Tezos_protocol_environment.Signature.Public_key_hash.t option

Originated contracts

type origination_nonce

Originated contracts handles are crafted from the hash of the operation that triggered their origination (and nothing else). As a single operation can trigger several originations, the corresponding handles are forged from a deterministic sequence of nonces, initialized with the hash of the operation.

val originated_contract : origination_nonce -> contract

val originated_contracts : since:origination_nonce -> until:origination_nonce -> contract list

val initial_origination_nonce : Tezos_protocol_environment.Operation_hash.t -> origination_nonce

val incr_origination_nonce : origination_nonce -> origination_nonce

val is_originated : contract -> Contract_hash.t option

Human readable notation

type Tezos_protocol_environment.Error_monad.error += Invalid_contract_notation of string

val to_b58check : contract -> string

val of_b58check : string -> contract Tezos_protocol_environment.Error_monad.tzresult

val pp : Tezos_protocol_environment.Format.formatter -> contract -> unit

val pp_short : Tezos_protocol_environment.Format.formatter -> contract -> unit

Serializers

val encoding : contract Tezos_protocol_environment.Data_encoding.t

val origination_nonce_encoding : origination_nonce Tezos_protocol_environment.Data_encoding.t

val rpc_arg : contract Tezos_protocol_environment.RPC_arg.arg

module Index : Storage_description.INDEX with type t = t