Module Vote.Proposals

Set of protocol proposal with corresponding proposer delegate

type context = Raw_context.t

type elt = Tezos_protocol_environment.Protocol_hash.t * Tezos_protocol_environment.Signature.Public_key_hash.t

The type of elements.

val mem : context -> elt -> bool Tezos_protocol_environment.Lwt.t

Tells if a elt is a member of the set

val add : context -> elt -> Raw_context.t Tezos_protocol_environment.Lwt.t

Adds a elt is a member of the set

val del : context -> elt -> Raw_context.t Tezos_protocol_environment.Lwt.t

Removes a elt of the set ; does nothing if not a member

val set : context -> elt -> bool -> Raw_context.t Tezos_protocol_environment.Lwt.t

Adds/Removes a elt of the set

val elements : context -> elt list Tezos_protocol_environment.Lwt.t

Returns the elements of the set, deserialized in a list in no particular order.

val fold : context -> init:'a -> f:(elt -> 'a -> 'a Tezos_protocol_environment.Lwt.t) -> 'a Tezos_protocol_environment.Lwt.t

Iterates over the elements of the set.

val clear : context -> Raw_context.t Tezos_protocol_environment.Lwt.t

Removes all elements in the set